Workaround for reference slash prefix in libgit2.
When libgit2 is passed a path to a working directory instead of a
git directory the names returned from git_reference_listall (and
perhaps other similar methods) will be prefixed with a slash such
that insteaf of refs/heads/master it'll return /refs/heads/master.

LibGitSharp always does it's string prefix comparisons without a
starting slash (which seems to be the correct thing to do).

Includes test which verifies the problem by copying the sample working
directory and performing the same test (CanListAllBranches) that's run
for the bare repository.

Update constructor documentation to reflect that it's possible
to pass the path to a working directory.