Fix up marshalling of strings returned by libgit2
Previously we were doing this which is not good:

return Marshal.PtrToStringAnsi(intPtr);

You can see the failing test that was added to ConfigurationFixture.cs to
demonstrate the problem (setting and getting a user name with unicode chars in
the config). This brought up the large problem of how we were dealing with
libgit2 methods that returned string values.

There was another subtle issue in the Utf8Marshaler where we were freeing
memory that we didn't own.