Add NuGet License
NuGet is licensed under Apache 2.0 therefore we should provide a
copy of the license alongside our binary redistribution of it in
our Libs folder.  Note that because we don't ship NuGet in the
binary version of the LibGit2Sharp we don't need to inlucde
a reference to it in our NOTICE file etc.