Test case to repro libgit2/libgit2sharp#387
When invalid encoding is in the author name or other commit metadata,
libgit2 always returns the text as if it where UTF8 encoded (which it
should be). jGit will attempt to parse using the current systems
default encoder. Core git will simply return the bytes leaving it
up to the client shell to interpret (which posh-git does by
defaulting to cp-1252 but Git Bash in msysgit shows differently)