Make the object database use posix-style paths.
Modify a test case to check whether the object database accepts paths that
are not valid on Windows.

Add in that same fixture a test of TreeDefinition.Remove .

Replace instances of Path.Combine by FilePath.CombineGitPaths() .

Change instances of FilePath to instances of string in the object
database, add some tests, add a static method to concatenate git paths.

Update methods documentation to explicitly say that the paths to use are
relative posix-style paths in TreeDefinition, Commit and Tree.