Make AppVeyor create NuGet packages upon a merge
Prior to generating the packages, the AssemblyInfo.cs is patched with
the build version. Additionally, the assembly informational version
will bear a `-pre` suffix would this build result from a push on a
branch which isn't `master`.

Produced packages will be published as build artifacts.

No packaging will be performed when the build has been triggered from a Pull Request.