Implement git log --follow
This commit basically implements the git log --follow <path> command. It adds
the following two methods to the IQueryableCommitLog interface:

IEnumerable<LogEntry> QueryBy(string path);
IEnumerable<LogEntry> QueryBy(string path, FollowFilter filter);

The corresponding implementations are added to the CommitLog class. The actual
functionality is implemented by the FileHistory class that is part of the
LibGit2Sharp.Core namespace.

Related to topics #893 and #89