Catch expected exceptions thrown by DirectoryHelper
The DirectoryHelper.DeleteDirectory() method throws exceptions.
Affected tests will be reported as FAILED although this happens
during cleanup after the test was run.

This commit deals with those exceptions by retrying to delete
the directory up to four times, which removes directories in all
but rare cases. The method catches DirectoryNotFoundException,
IOException, and UnauthorizedAccessException and will only log a
trace message if it can't finally delete the directory. It will
rethrow any other type of exception.

Closes #970.