Make NuGet package copy native binaries on windows only
The props file added to the NuGet package in #821 makes sure the Windows
native binaries are copied to the correct location. However, as pointed out
in #835, this now happens even when the package is installed on a non-Windows
platform.

This can be prevented by adding a condition to the imported ItemGroup that
makes it be ignored when running on a non-Windows platform.

Closes #835