PoC for writing directly into the next stream
The filters are run in-line with the rest of the system, and libgit2 is
set up to have its last filter/writestream write to the
filesystem. Instead of using an intermediate buffer/file/pipe, we can
make the writes from the user's Smude() or Clean() filter into their output
stream propagate immediately into the next filter, leading up the chain
to the filesystem, removing any intermediate buffering other than that
which is necessary for the filters themselves to work.

In the general case, we only have the one clean or smudge filter, which
means that even in the case of out-of-band storage for large files,
where a small blob can cause very large outputs, we would still write
directly to the filesystem.

by J Wyman (@whoisj): Buffered writes to next filter to minimize managed to native marshaling for performance optimizations

by J Wyman (@whoisj): Split `WriteStream` into its own file and cleaned it up (checks, style, etc)