Remove union merge driver for sln, csproj files
The union merge driver is not appropriate for sln or csproj
files.  It is useful (as described by Junio) to "merge two
shopping lists".

This will work in trivial examples for sln and csproj files,
for example when one branch adds a new file and the other branch
does too, but it will often silently produce corrupt output.