Explicitly set xUnit MSBuild WorkingFolder property
It seems that the xUnit.net 2.0 MSBuild test runner does not play nicely
with mono 3.12.1, which is being used on the linux Travis CI builds.

Setting this property appears to stop the intermittent failures by
preventing xUnit from forgetting how to find the Resources folder in the
middle of running the tests.