Explore moving the command-simulating methods to their own namespace
A lot of the methods on Repository are actually trying to replicate git
commands, which makes it harder to see what's the operation vs command.

Move them to their own namespace to make it clear what they're trying to
achieve. We start with a few forms of the git-checkout command.