Disentangle the reference and config iterators
While these both work roughly in the same manner (as they both represent
iterators) the actual code duplication between these two is rather small
as each `_next()` method is different and how we return the managed values
is different. The net reduction in lines of code indicates that this is
indeed the case.