Use a common abstract class for the handles
Instead of copying each and every class, we can put the common
functionality into a single class, using a void pointer and use the
template to generate the bits which need to know about the pointer we're
dealing with.

Freeing and the implicit conversion to the pointer type are thus still
there, but we now don't copy the disposing code everywhere.