Remove the code we have just for mocking
We make the constructors internal but then make sure we cave
constructors with zero arity so mocking them via e.g. Moq is
easier. With this we're making out code harder to mock, but then add
ways to add another layer of mocking easier.

We enforce this constructors for mocking in the tests, without regard
for whether they make sense or not. Often it would make a lot more sense
to expose a public constructor for types so an application can return
fake results.