Remove the last uses of explicit config paths for Repository
These tests use the option to set a global configuration file with the
filemode we want. But the filemode setting should be set per-repository
as it comes down to the workdir for each repository.

Switch to setting the configuration in the local configuration, which is
also more in line with how it would exist in the wild.