Don't load an object just to check its ID
In order to check for equality we just need to compare the ID of the
object. We do not need to look up the object but can compare their
identifiers directly.

This also requires us to modify Branch's check for the current branch,
but the new code more accurately reflects the check we do want to
perform. Namely whether HEAD points to our reference name.