Don't provide a diff notify callback unless necessary
We only seem to care about what ends up being collected in the
`MatchedPathsAggregator` when we've got some file paths to match agains
and either an `OnUnmatchedPath` delegate or instructions to throw
on unmatched paths.

If none of those condition are true we shouldn't be sending a callback
function pointer at all and by not doing so we avoid one marshalled invocation
per delta in the diff.