Add a RemoteName property to avoid lookups
Sometimes all you need is the name and right now there's no way of getting
it short of git_remote_lookup. While going through a lookup guarantees
that the remote exists that shouldn't be the concern of the Branch, it
should just let the caller know what's been configured as it's remote.