Don't use FilePathMarshaller when passing Git paths
We treat a lot of the data as paths which would be available on the host
system, but this doesn't need to be the case. There are paths which the
filesystem or .NET won't accept which work just fine as paths inside
Git.

This also means the return values are also marshalled as strings, since
we can't assume they contain paths which the path marshaller would
like. It also keeps the separator as the slash rather than convert it
into something which seems native, but is different from what we'd
accept.

The paths which we do refer to as paths inside the workdir are still
marhshalled as paths.