NativePath: use Assembly.CodeBase
Use `Assembly.CodeBase` instead of `Assembly.EscapedCodeBase` since
the latter does not actually properly escape the path in a reversible
way.

eg, if a path contains a space, it will be encoded as `%20` in the
`EscapedCodeBase` however if a path contains a literal `%20` then it
will _also_ be encoded as `%20`.