Update test comments and remove unnecessary check
We have exactly two optional regex matching groups. Regex itself guarantees that the required groups were present, so the last half of this test that confirms that a hard-coded set of these groups were present is not testing libgit2sharp as much as it is testing regex itself, which is difficult to maintain and pointless, so I removed it rather than fixing it when I added the second optional group.