nuget: don't try to find msbuild (incorrectly)
With the changes in the way MSBuild is bundled and installed in Visual
Studio 2015+, we can no longer use the old hacky way to find where
msbuild is installed.

This busted AppVeyor whose image has VS 2015 installed and an old
(incompatible) version of msbuild installed elsewhere.  We would find
only the old version.

Require nuget package users to have msbuild in their path.