Exceptions: refactor native exceptions to include code
Include the error code as a first-class citizen of exceptions that
correspond to libgit2 errors.  Previously, we would store the error
code in exception data when creating the exception from a libgit2 native
error.  However, we may also want to _generate_ these exceptions and
have them converted to a libgit2 error.  (For example, a callback may
want to return GIT_EUSER, and so throw a `UserCancelledException`.)

Therefore, provide exceptions with knowledge of their corresponding
libgit2 error code so that we can round-trip that data.