Use existing repository configuration
Previously we rebuilt configuration for repository on Configuration
object creation. It leaded to the issues as sometimes actual git config
differed from what we build. That for instance happened with worktrees,
when local config path was calculated differently.
Changed logic to open the existing configuration if repository exists.
This way we are always consistent with libgit2 behavior.