Move credential caching to ManagedHttpSmartSubtransport, no longer static.
When making multiple requests to the same repository URL with different credentials, the first set of successful credentials would be reused indefinitely (as the credential cache is static). This can cause an unexpected re-use of credentials when the same URL is passed to seemingly independent git operations. This change creates a new HttpClientHandler and CredentialHandler for each ManagedHttpSmartSubtransport to eliminate any sharing of credentials between operations.

It's possible that this could be a breaking change for some users; if a consumer is only setting credentials for the first operation, subsequent operations will now fail.

This change will result in more calls to the credential provider. If the provider is doing complex authentication logic (potentially calling external services), they will need to be responsible for caching between operations.